refactor: move verifyFormFields to separate utils file
