+++ app/components/Form/ApplicationForm.tsx
@@ -12,6 +12,7 @@ import { useSubmitApplicationMutation } from 'schema/mutations/application/submi
 import { acknowledgementsEnum } from 'formSchema/pages/acknowledgements';
 import { updateApplicationFormMutation } from '__generated__/updateApplicationFormMutation.graphql';
 import { useUpdateApplicationForm } from 'schema/mutations/application/updateApplicationForm';
+import verifyFormFields from 'utils/verifyFormFields';
 import { ReviewField } from 'components/Review';
 import SubmitButtons from './SubmitButtons';
 import FormBase from './FormBase';
@@ -88,28 +89,6 @@ interface SubmissionFieldsJSON {
   submissionTitle?: string;
 }
 
-const verifyFormFields = (formSectionData, formSectionName) => {
-  // verify that the form fields are in the seciton of the schema we are saving
-  // to prevent the bug where fields were being saved in the wrong section
-  const newFormFieldNames = formSectionData && Object.keys(formSectionData);
-  const formSchemaSectionFieldNames = Object.keys(
-    schema.properties[formSectionName]['properties']
-  );
-  const incorrectFormFields = newFormFieldNames?.filter(
-    (fieldName) => !formSchemaSectionFieldNames.includes(fieldName)
-  );
-
-  const verifiedFormSectionData =
-    formSectionData &&
-    Object.fromEntries(
-      Object.entries(formSectionData).filter(
-        ([key]) => !incorrectFormFields.includes(key)
-      )
-    );
-
-  return verifiedFormSectionData;
-};
-
 export const mergeFormSectionData = (
   formData,
   formSectionName,
@@ -126,8 +105,12 @@ export const mergeFormSectionData = (
   } else if (formData[formSectionName]) {
     newFormData = { ...formData };
     newFormData[formSectionName] = {
-      ...verifyFormFields(newFormData[formSectionName], formSectionName),
-      ...verifyFormFields(calculatedSection, formSectionName),
+      ...verifyFormFields(
+        newFormData[formSectionName],
+        formSectionName,
+        schema
+      ),
+      ...verifyFormFields(calculatedSection, formSectionName, schema),
     };
   } else {
     newFormData = { ...formData };
